+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v3.0.1
 * Changed: Find Path to not return a path that contains one or more vertices that is can't be retrieved because of visibility restrictions
 * Added: Added a hasId method to the Query class to allow searches to be filtered by element ID.
+* Fix: Extended data element type value for edges
 
 # v3.0.0
 * Changed: Removed ES 2 support and replaced it with ES 5 support